requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.12.0,<3.0.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0

